allow sending key/value pairs with dispatch
Make pulumictl dispatch more generally useful, and allow key/value
pairs in addition to a single ref.  For starters we'll use this to
dispatch a GitHub actions job that will update the
`latest-dev-version` file (see https://github.com/pulumi/docs/pull/10365),
but this can be used for other workflows as well.